src/fastapi-order-service/requirements.txt
fastapi==0.65.3
requests==2.25.1
uvicorn==0.14.0
tortoise-orm==0.17.4
databases[postgresql]==0.2.6
python-redis-cache==1.2.0
redis==4.5.4
pika==1.3.1
coloredlogs==15.0.1